Simplification of the blockfrostRequest
I believe I have identified a couple of things to improve the `blockfrostRequest` function in `src/api/util.js`:

- While this function is using promises, it's already an asynchronous function.
- This function seems to be repeating the code used to build the outgoing request.
- This function is also using `setInterval` to loop re-building the outgoing request.

`setInterval` in particular can make the code hard to follow. Here I am introducing a utility function `delay` which allows us to delay the iteration of loops inside asynchronous functions.

So, this PR proposes a refactoring of the `blockfrostRequest` function to:

- Build the outgoing request repeatedly with a single loop block.
- Delay the loop block using the new `delay` function.
- Thus, removing all references for `Promise` inside this asynchronous function.

The tests keep passing 🙂
How does this look? Feedback appreciated.