chore: initial workflow for issue moderation
This workflow will trigger when comments are left on issues, or existing comments are edited.
It has some basic keywords that can be expanded on, but the purpose of this change is to see
the workflow running, with the results printed to the console. After this is proven, additional
steps will be made to use the information effectively.